Revert "Revert: Make mkosi a python module and generate script via entrypoint"
This reverts commit 7b86ed98f07d20a46aaeaa64db03499b443412e9.
